Move App and Provider in components
