+++ site/source/components/App.tsx
@@ -19,26 +19,26 @@ import {
 import { useIsEmbedded } from '@/components/utils/useIsEmbedded'
 import { Container, Spacing } from '@/design-system/layout'
 
+import { useAxeCoreAnalysis } from '../hooks/useAxeCoreAnalysis'
+import { useGetFullURL } from '../hooks/useGetFullURL'
+import { useSaveAndRestoreScrollPosition } from '../hooks/useSaveAndRestoreScrollPosition'
+import Accessibilité from '../pages/Accessibilité'
+import Budget from '../pages/Budget/Budget'
+import Créer from '../pages/Creer'
+import IntegrationTest from '../pages/Dev/IntegrationTest'
+import Documentation from '../pages/Documentation'
+import Iframes from '../pages/Iframes'
+import Landing from '../pages/Landing/Landing'
+import Nouveautés from '../pages/Nouveautes/Nouveautes'
+import Offline from '../pages/Offline'
+import Plan from '../pages/Plan'
+import Simulateurs from '../pages/Simulateurs'
+import Stats from '../pages/Stats/LazyStats'
+import Assistants from '../pages/assistants'
+import Gérer from '../pages/gerer'
+import Integration from '../pages/integration/index'
+import { useSitePaths } from '../sitePaths'
 import Provider, { ProviderProps } from './Provider'
-import { useAxeCoreAnalysis } from './hooks/useAxeCoreAnalysis'
-import { useGetFullURL } from './hooks/useGetFullURL'
-import { useSaveAndRestoreScrollPosition } from './hooks/useSaveAndRestoreScrollPosition'
-import Accessibilité from './pages/Accessibilité'
-import Budget from './pages/Budget/Budget'
-import Créer from './pages/Creer'
-import IntegrationTest from './pages/Dev/IntegrationTest'
-import Documentation from './pages/Documentation'
-import Iframes from './pages/Iframes'
-import Landing from './pages/Landing/Landing'
-import Nouveautés from './pages/Nouveautes/Nouveautes'
-import Offline from './pages/Offline'
-import Plan from './pages/Plan'
-import Simulateurs from './pages/Simulateurs'
-import Stats from './pages/Stats/LazyStats'
-import Assistants from './pages/assistants'
-import Gérer from './pages/gerer'
-import Integration from './pages/integration/index'
-import { useSitePaths } from './sitePaths'
 
 type RootProps = {
 	basename: ProviderProps['basename']
