+++ site/source/components/Provider.tsx
@@ -18,14 +18,14 @@ import { Link } from '@/design-system/typography/link'
 import { Body, Intro } from '@/design-system/typography/paragraphs'
 import { useIframeResizer } from '@/hooks/useIframeResizer'
 
-import { TrackingContext } from './components/ATInternetTracking'
-import { createTracker } from './components/ATInternetTracking/Tracker'
-import { ServiceWorker } from './components/ServiceWorker'
-import { DarkModeProvider } from './contexts/DarkModeContext'
-import { Message } from './design-system'
-import * as safeLocalStorage from './storage/safeLocalStorage'
-import { store } from './store/store'
-import { inIframe } from './utils'
+import { DarkModeProvider } from '../contexts/DarkModeContext'
+import { Message } from '../design-system'
+import * as safeLocalStorage from '../storage/safeLocalStorage'
+import { store } from '../store/store'
+import { inIframe } from '../utils'
+import { TrackingContext } from './ATInternetTracking'
+import { createTracker } from './ATInternetTracking/Tracker'
+import { ServiceWorker } from './ServiceWorker'
 
 type SiteName = 'mon-entreprise' | 'infrance' | 'publicodes'
